refactor(icons): flatten dist directory
