+++ packages/icons/package-scripts.js
@@ -1,6 +1,10 @@
 const getScripts = require("@ui5/webcomponents-tools/icons-collection/nps.js");
 
-const scripts = getScripts();
+const options = {
+	collectionName: "SAP-icons",
+};
+
+const scripts = getScripts(options);
 
 module.exports = {
 	scripts
