+++ packages/icons/src/json-imports/Icons.js
@@ -1,5 +1,5 @@
 import { registerIconBundle } from "@ui5/webcomponents-base/dist/asset-registries/Icons.js";
 
-import SAPIcons from "../generated/assets/icon-collections/SAP-icons.json";
+import SAPIcons from "../generated/assets/SAP-icons.json";
 
 registerIconBundle("SAP-icons", SAPIcons);
