+++ packages/main/src/DateTimePicker.js
@@ -2,7 +2,7 @@ import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.j
 import getLocale from "@ui5/webcomponents-base/dist/locale/getLocale.js";
 import LocaleData from "@ui5/webcomponents-localization/dist/LocaleData.js";
 import CalendarDate from "@ui5/webcomponents-localization/dist/dates/CalendarDate.js";
-import "@ui5/webcomponents-icons/dist/icons/date-time.js";
+import "@ui5/webcomponents-icons/dist/date-time.js";
 import {
 	isLeft,
 	isRight,
