+++ packages/main/src/TimePicker.js
@@ -21,7 +21,7 @@ import {
 	isPageUpShiftCtrl,
 	isPageDownShiftCtrl,
 } from "@ui5/webcomponents-base/dist/Keys.js";
-import "@ui5/webcomponents-icons/dist/icons/time-entry-request.js";
+import "@ui5/webcomponents-icons/dist/time-entry-request.js";
 import Icon from "./Icon.js";
 import PopoverHorizontalAlign from "./types/PopoverHorizontalAlign.js";
 import ResponsivePopover from "./ResponsivePopover.js";
